Move JNI code to JNITunnelController
Much better way to avoid sharing state.
